requirements.txt
aiofiles==0.5.0
aniso8601==7.0.0
async-exit-stack==1.0.1
async-generator==1.10
attrs==19.3.0
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
dataclasses==0.7
dnspython==2.0.0
email-validator==1.1.1
fastapi==0.59.0
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.0.4
h11==0.9.0
httptools==0.1.1
idna==2.10
importlib-metadata==1.7.0
iniconfig==1.0.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
more-itertools==8.4.0
multidict==4.7.6
orjson==3.2.2
packaging==20.4
pkg-resources==0.0.0
pluggy==0.13.1
promise==2.3
py==1.9.0
pydantic==1.6.1
pyparsing==2.4.7
pytest==6.0.1
python-multipart==0.0.5
PyYAML==5.3.1
requests==2.24.0
Rx==1.6.1
six==1.15.0
soupsieve==2.0.1
starlette==0.13.4
toml==0.10.1
typing-extensions==3.7.4.2
ujson==3.0.0
urllib3==1.25.9
uvicorn==0.11.6
uvloop==0.14.0
vcrpy==4.1.0
websockets==8.1
wrapt==1.12.1
yarl==1.5.1
zipp==3.1.0
